Functionalized most of the runit loop, adjusted signal handling.
Most of the signal control has been removed or replaced by asynchronous
signal handling. (asynchronous to the main loop, not a separate process)

Added a watchlist for spawned children so process control can be a
little more loose. My goal is to make it easier to add new functionality
to the runit loop.